requirements.txt
Django==5.2.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dj-database-url==2.2.0
Pillow==11.3.0
python-dateutil==2.9.0.post0